feat: Modify the stage name to be optional
Remove the requirement for the stage name to be mandatory and use the
general max length validation instead (CU#85zt70f4b). Also, default to
stage name if available for header displays, profile page and sidebar,
otherwise use the first name and last name which are required
(85zt01b7e).