feat: Update support links for artist testing
Update the support link for ask the community to the discord channel and
adds discord logo to the assets folder to substitute the telegram logo,
per CU-86a0yf03x. Also update the support link for email support to
point to the ClickUp submission form page, per CU-86a10trb9.