feat: Add payment types and release summary dialog
This commit introduces a new payment type option to the song upload
and edit flows. It also adds a Release Summary Dialog to provide
users with a detailed breakdown of costs and payment methods.
Additionally, it includes a new CheckCircleRadioIcon for UI
enhancements and updates the theme to support active background
gradients.